dom/bindings/test/test_sequence_detection.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1066432
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1066432</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">
  /* global TestInterfaceJS */
  /** Test for Bug 1066432 **/
  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({set: [["dom.expose_test_interfaces", true]]}, function() {
    var testInterfaceJS = new TestInterfaceJS();
    ok(testInterfaceJS, "got a TestInterfaceJS object");

    var nonIterableObject = {[Symbol.iterator]: 5};

    try {
      testInterfaceJS.testSequenceOverload(nonIterableObject);
      ok(false, "Should have thrown in the overload case"); // see long comment above!
    } catch (e) {
      is(e.name, "TypeError", "Should get a TypeError for the overload case");
      ok(e.message.includes("not iterable"),
         "Should have a message about being non-iterable in the overload case");
    }

    try {
      testInterfaceJS.testSequenceUnion(nonIterableObject);
      ok(false, "Should have thrown in the union case");
    } catch (e) {
      is(e.name, "TypeError", "Should get a TypeError for the union case");
      ok(e.message.includes("not iterable"),
         "Should have a message about being non-iterable in the union case");
    }

    SimpleTest.finish();
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1066432">Mozilla Bug 1066432</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
